refactor: move controller to reasonable package name
