+++ graphene-reader/src/main/kotlin/com/graphene/reader/controller/graphene/PathsController.kt
@@ -1,4 +1,4 @@
-package com.graphene.reader.controller
+package com.graphene.reader.controller.graphene
 
 import net.iponweb.disthene.reader.service.index.ElasticsearchIndexService
 import net.iponweb.disthene.reader.service.stats.StatsService
