+++ graphene-reader/src/main/kotlin/com/graphene/reader/controller/graphite/request/MetricsFindRequest.kt
@@ -1,0 +1,7 @@
+package com.graphene.reader.controller.graphite.request
+
+data class MetricsFindRequest(
+  val query: String,
+  val from: String?,
+  val to: String?
+)
